Include workaround for weird install hack in makefiles/bmv2.mk
The way the code is set up, it expects the configure script to be called
with --prefix=/bmv2_pd, which doesn't work with Yocto.  The workaround
is to pass the "correct" values for the variables it uses to decide
where to install stuff (includedir, libdir).

The reason why passing prefix doesn't work is because when all this
makes it to the Makefile, the variables no longer reference $(prefix),
but instead they have hardcoded values derived from the initial prefix
setting.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@hpe.com>